+++ packages/orq-rc/RELEASES.md
@@ -739,3 +739,13 @@ Based on:
 - [python v3.2.0-rc.36] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.36] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.36 - packages/orq-rc
+
+## 2025-03-10 14:51:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.513.6 (2.545.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.37] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.37] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.37 - packages/orq-rc